Convert Settings test to tlr.
